+++ src/_spec/fixes-4.spec.ts
@@ -61,12 +61,12 @@ describe("fix-3", function() {
         const schemas = [compile(`
             interface A {
                 @constraint('unique', ['p', 'r'])
-                a: {p: string, q: string, r: string}[];
+                a: {p: string | null, q: string, r: string}[];
             }
         `), compile(`
             interface A {
                 @constraint('unique', ['p', 'r'])
-                a?: {p: string, q: string, r: string}[];
+                a?: {p: string | null, q: string, r: string}[];
             }
         `)];
         for (const schema of schemas) {
@@ -108,6 +108,101 @@ describe("fix-3", function() {
                     {p: '1', q: '4', r: '3'},
                 ]}, ty, ctx);
                 expect(z).toEqual(null);
+            }
+            {
+                const ctx: Partial<ValidationContext> = {
+                    checkAll: true,
+                    stereotypes: new Map([
+                        ...dateStereotypes,
+                    ]),
+                    customConstraints: new Map([
+                        ...uniqueConstraints,
+                    ]),
+                };
+                const z = validate<any>({a: [
+                    {p: '1', q: '2', r: '3'},
+                    {p: '2', q: '3', r: '4'},
+                    {p: null, q: '4', r: '5'},
+                    {p: null, q: '4', r: '5'},
+                ]}, ty, ctx);
+                expect(z).toEqual(null);
+            }
+        }
+    });
+    it("unique-3", function() {
+        const schemas = [compile(`
+            interface A {
+                @constraint('unique-non-null', ['p', 'r'])
+                a: {p: string | null, q: string, r: string}[];
+            }
+        `), compile(`
+            interface A {
+                @constraint('unique-non-null', ['p', 'r'])
+                a?: {p: string | null, q: string, r: string}[];
+            }
+        `)];
+        for (const schema of schemas) {
+            const ty = getType(schema, 'A');
+            {
+                const ctx: Partial<ValidationContext> = {
+                    checkAll: true,
+                    stereotypes: new Map([
+                        ...dateStereotypes,
+                    ]),
+                    customConstraints: new Map([
+                        ...uniqueConstraints,
+                    ]),
+                };
+                const z = validate<any>({a: [
+                    {p: '1', q: '2', r: '3'},
+                    {p: '2', q: '3', r: '4'},
+                    {p: '3', q: '4', r: '5'},
+                ]}, ty, ctx);
+                expect(z).toEqual({value: {a: [
+                    {p: '1', q: '2', r: '3'},
+                    {p: '2', q: '3', r: '4'},
+                    {p: '3', q: '4', r: '5'},
+                ]}});
+            }
+            {
+                const ctx: Partial<ValidationContext> = {
+                    checkAll: true,
+                    stereotypes: new Map([
+                        ...dateStereotypes,
+                    ]),
+                    customConstraints: new Map([
+                        ...uniqueConstraints,
+                    ]),
+                };
+                const z = validate<any>({a: [
+                    {p: '1', q: '2', r: '3'},
+                    {p: '2', q: '3', r: '4'},
+                    {p: '1', q: '4', r: '3'},
+                ]}, ty, ctx);
+                expect(z).toEqual(null);
+            }
+            {
+                const ctx: Partial<ValidationContext> = {
+                    checkAll: true,
+                    stereotypes: new Map([
+                        ...dateStereotypes,
+                    ]),
+                    customConstraints: new Map([
+                        ...uniqueConstraints,
+                    ]),
+                };
+                const z = validate<any>({a: [
+                    {p: '1', q: '2', r: '3'},
+                    {p: '2', q: '3', r: '4'},
+                    {p: null, q: '4', r: '5'},
+                    {p: null, q: '4', r: '5'},
+                ]}, ty, ctx);
+                expect(z).toEqual({value: {a: [
+                    {p: '1', q: '2', r: '3'},
+                    {p: '2', q: '3', r: '4'},
+                    {p: null, q: '4', r: '5'},
+                    {p: null, q: '4', r: '5'},
+                ]}});
             }
         }
     });